Refector/rename mention view
