+++ lib/prompt-editor/src/v2/promptInput-react.ts
@@ -5,7 +5,7 @@ import { Item } from "./Suggestions"
 import { ActorRefFrom, fromCallback } from "xstate"
 import { useEffect, useMemo, useRef } from "react"
 import { Node } from "prosemirror-model"
-import { MentionView } from "./mentionNode"
+import { MentionView } from "./MentionView"
 import { replaceDocument } from "./prosemirror-utils"
 import { EditorState } from "prosemirror-state"
 import { Position } from "./atMention"
